Add project create and update mixins to dashboard.
